migrate ProReportsReadReminder
