+++ app/services/emails/EmailDefinitionsPro.scala
@@ -13,6 +13,8 @@ import utils.FrontRoute
 import utils.SIREN
 
 import java.net.URI
+import java.time.OffsetDateTime
+import java.time.Period
 
 object EmailDefinitionsPro {
 
@@ -165,6 +167,33 @@ object EmailDefinitionsPro {
         override def getBody: (FrontRoute, EmailAddress) => String =
           (frontRoute, _) => views.html.mails.professional.reportReOpening(report)(frontRoute).toString
       }
+  }
+
+  case object ProReportsReadReminder extends EmailDefinition {
+    override val category = Pro
+
+    override def examples = {
+      val report1 = genReport
+      val report2 = genReport.copy(companyId = report1.companyId)
+      val report3 = genReport.copy(companyId = report1.companyId, expirationDate = OffsetDateTime.now().plusDays(5))
+      Seq(
+        "reports_transmitted_reminder" -> (recipient =>
+          EmailImpl(List(recipient), List(report1, report2, report3), Period.ofDays(7))
+        )
+      )
+    }
+
+    case class EmailImpl(
+        recipients: List[EmailAddress],
+        reports: List[Report],
+        period: Period
+    ) extends ProFilteredEmailMultipleReport {
+      override val subject: String = EmailSubjects.REPORT_TRANSMITTED_REMINDER
+
+      override def getBody: (FrontRoute, EmailAddress) => String =
+        (frontRoute, _) =>
+          views.html.mails.professional.reportsTransmittedReminder(reports, period)(frontRoute).toString
+    }
   }
 
 }
