+++ app/tasks/report/ReportRemindersTask.scala
@@ -12,8 +12,8 @@ import play.api.Logger
 import repositories.event.EventRepositoryInterface
 import repositories.report.ReportRepositoryInterface
 import repositories.tasklock.TaskRepositoryInterface
-import services.emails.Email.ProReportsReadReminder
 import services.emails.Email.ProReportsUnreadReminder
+import services.emails.EmailDefinitionsPro.ProReportsReadReminder
 import services.emails.Email
 import services.emails.MailServiceInterface
 import tasks.ScheduledTask
@@ -21,16 +21,15 @@ import tasks.getTodayAtStartOfDayParis
 import utils.Constants.ActionEvent._
 import utils.Constants.EventType.SYSTEM
 import utils.EmailAddress
+import utils.Logs.RichLogger
 
 import java.time._
 import java.util.UUID
 import scala.concurrent.ExecutionContext
 import scala.concurrent.Future
+import scala.concurrent.duration.FiniteDuration
 import scala.util.Failure
 import scala.util.Success
-import utils.Logs.RichLogger
-
-import scala.concurrent.duration.FiniteDuration
 class ReportRemindersTask(
     actorSystem: ActorSystem,
     reportRepository: ReportRepositoryInterface,
@@ -97,7 +96,7 @@ class ReportRemindersTask(
               readByProsSent <- sendReminderEmailIfAtLeastOneReport(
                 readByPros,
                 users,
-                ProReportsReadReminder,
+                ProReportsReadReminder.EmailImpl,
                 EMAIL_PRO_REMIND_NO_ACTION
               )
               notReadByProsSent <- sendReminderEmailIfAtLeastOneReport(
